v 1.3 Sliders and Offsets
I couldn't get the offset working quite right. the sliders are calling the update value a lot when the slider is touched. I'm reading the non modified values of the trail and writing them back into the new vector, but there's some sort of rounding error I think that's causing the non modified values to slowly drift. I didn't have this problem in my other method where I combined the vertrical and horizontal offsets into one method. I might just go back to that to fix things.
I reimplmented the slider offset as one function and now I'm not getting the rounding errors. It seems to work for moving the offset, but the trails need to be redrawn with the offset now, which is a bit beyond what I expected to do to get this workin