Marker Persistance and Trail Toggle
The markers were dissapearing due to their parent object being associated with the hand gestures. When the hand with the spawn menu on it goes away so do all the markers. I updated the marker spawning script to take the position of the marker spawn from a Game object transform. I used a child of the camera to get it's coordinates and offset it by 1 unit in Z. Then I spawned the marker with a parent object a 0,0,0 to keep track of the markers and persist them when the hands get deactivated.
I also hooked up the trails toggle to the headsets line renderers, so they can be turned on and off.