Editor Crash in QR
I tracked down the crashing of the editor to the QR object in Unity. Then I went into the script to shut down the services I could find via OnApplicationQuit. When playmode is exited this is called and can be used to close the program. I tried calling the watcher stop, but the crash seems to be in the QRCodeWatcher.RequestAccessAsync().Result. call
When those are commented out the editor can be re-played without crashing. I don't quite know how to handle that async call, but we've found the source.
I've just been disabling the QR gameobject in the Unity editor for other testing for now. Cheers!