Added 2 new text indicators
Added new text indicator for both left and right palms to inform the users that the palms should be oriented toward their face in order to see the map and the menu --> improving usability and experience. Note: Still need to resize the font after testing and maybe need to adjust the color as well.