Adding code to hide the text prompts
When the user turns their palm away from them the instruction text will show up. When the palm is towards the users, the text will disappear. We currently don't have it showing up on the first view, but it will show up after the hand is actiavted and deactivated. We're trying to fix that.