Adding a script MapText for Displaying Location Text for Map
Added v1.7 and worked on attaching a script for the Map_Text object. Right now, there are still some problem with using GetWebRequest to work properly, so have to fix that, which I'm not sure why the request fail. Also contacted Lance about the issue w/ getting the name of the image file. So, need to fix that as well as soon as I hear back from Lance.