Markers are posting! yay
We got the markers to post to the server! The style class needs to be implemented and our time is weird, but we'll need to check how the server time can be used in Unity. We had a little confusion about the transormation and pose of the markers, but we found that was because we were looking at the pose of the template object in the instantiate funciton instead of the new cloned object. once we got th reference to that the pose is now correct :)

Co-Authored-By: arishuang10 <71896453+arishuang10@users.noreply.github.com>