Hololens Button Menu Locked
We tried to use the near menu to allow for selection of the different features, but there's something wrong with the location tracking. We tried to override the positions to use the hand, but it's staying locked to the origin. We're probably going to use the hand menu in the future, but it's confusing as to why we can't get this to work.

Co-Authored-By: Pallasite <Pallasite@users.noreply.github.com>
Co-Authored-By: arishuang10 <71896453+arishuang10@users.noreply.github.com>