v1.9 - Rendering Featuring Fixing Few Structure
Add and fix a couple of implementation with regards to the feature. Right now, everything does not work. However, that is because we have a new set up for spawning the feature. The feature is no longer list but rather it's a dictionary. So, once we have that fix on Tuesday, everything theoretically should be able to work.