v1.9 updating the functionalities.
The front end is going to be different from the original spawnlist script, therefore I changed a couple things to make sure that everything else is working. Right now, the post functionality is the only one that I'm not sure if it works, because (can only know if it will after the front end part for spawning the markers are working.