Marker menu update
We've added the marker menu spawner with the string based key. hooked up the buttons to the correct string calls. Now we need to figure out why the menu doesn't track with the hand. It might have something to do with world locking before the hand is detected, but we're going to create a sandbox to try and understand the hand tracking. The buttons will spawn certain objects, but we're getting a JSON formatting error right now on spawn.

Co-Authored-By: arishuang10 <71896453+arishuang10@users.noreply.github.com>