v2.0 - Display all the feature is working properly
I combined the SpawnListIndex.cs script with the FeatureManager.cs script, so everything is now in FeatureManager.cs. When you start the scene right when you hit play, a;; the features from the server will be displayed automatically and will get updated throughout. The displaying of all feature is now integrated into the ListFeatures(). So, the previous DisplayAllFeatureGameObjects() found in the SpawnListIndex.cs no longer exists since it's substituted as mentioned previously. I've also added another method called, DeleteAll() in the FeatureManager.cs to help delete all the feature from the servers.