v2.3 - changed some stuffs after the last update
The text for the distance is still not working, however you can find the overall idea and code for the display within FeatureManager.cs.  Currently, the way that I'm calculating the distance is taking the x and z axes from the headset and feature positions. The headset position right now I've set it to the main camera position, which I guess does make sense. On the other hand, there are still some errors with float to string and mathematical variable conversion. The biggest problem currently is that when I ran the project, the TextMesh Pro variable (distance_text) always got set to null, which I can't seem to figure out why since I have made a prefab called FeatureText, which lives as the child of the "Distance Spawn Parent" empty game object (see the inspector), and dropped in as the reference within the "HandMenu_Large...." object under the FeatureManager.cs Script's property.