v2.3 - more minor changes
see the previous update for more explanation. For this update, I fixed the mathematical conversion errors, now it's only the error with the TestMesh Pro null problem and for some reason the text objects that were being spawned did not became the child of the "Distance Spawn Parent". I don't know why the text isn't instantiating.