Added Billboarding Effect
All the feature markers will now face the user wherever he or she goes. It's working correctly. Also, the distance text hierarchy has been further tested and working correctly as well. One caveat, though, is that somehow we are not locally instantiating the features listed from the server. As of now, we are only instantiating markers that we placed locally. This might be something that we would have to look into further.