Unity Input for Map
Added a unity input action for the map toggle and a script to handle a list of objects that need to be activated or deactivated. I don't see the other input assets, so I may have to combine that with what we already have. The map is head tracked now, but I'm not sure if it's hooked up correctly or in a good place.