Got the color of the float map to align + added one more prefab
The floating map icons' colors are now in alignment with the features created on the server. I added a feature + map icon prefab called 'danger' for the input 'm' button, but the scale is not quite right so it's still biohazard for now. However, I realized that even if the map is showing, the line rendering the map would not show up unless we trigger the PalmUpMap... this is kind of weird...