Navigation Update
Still working on the Navigation.cs script, I added a tag ('Icon') to each of the map icons for detecting if the icon is being pressed on the palm map. However, for some reason gameobject detection is not working (perhaps this functionality is disabled?). I also added a line renderer attached to the main camera. This is to draw a line between the user and the desire landmark/icon, but this is not working so far as well.