Buttons on the Map icons
We switched the interactions type to buttons instead of manipulatable. The collider boxes are a bit too big at the moment, but it's detecting on touch now instead of on click. The on click requires you to complete the press animation, but on touch just relies on the collider. Because of the smaller scale of the buttons it's difficult to get the on click to register in the Unity editor, but it might work in the HMD. We've left it for now as on touch.