Added Querying Server Functionality
The FindPath() in the Navigation.cs has this functionality semi-working (it's getting the right result, but there are some issues). However, the JSON format listed on the service (the point data) is not colon separated, which I cannot properly extract the data with this format inconsistency. So, I'll ask Lance about this.