Remove DLL Conflict
There seems to be a problem with the WindowsRT package that was installed vi a NuGet I think. The MRTK Tool doesn't detect it and isn't able to update it properly. We should either remove that from the project from the NuGet side of things or update it form there. I removed the duplicate package from the MRTK tool to rid us of the compiler error, but there's still another namespace problem.