Scene Understanding A GOGO
Eliminated the DLL conflict by removing the NuGet package and updating using the MRTK Tool. Added the updated versions of the MRTK plugins. Up from 2.7 to 2.8.
Imported the exapmple scene for the Scene understanding for Hunter to start integrating their work into.
Changed the .NET to .NET framwork in the player settings
REMOVED the ENABLE_WINMD_SUPPORT script compliation directive because it was causing namespace errors. I believe it was forcing support even when on the wrong platform, eg. the unity editor. I don't know if that broke anything, but it did fix errors!