Performance Update Distance Calc
I added some logical control to elimniate the repeated insantiation of the headset map icons. Now they will only be instantiated once and the resulting game object will have it's properties update accordingly. This saved about 15-20ms according to the deeep profiler. There is still some tex mesh pro updates that are kind of expensive, but I'm not sure if that will be an issue.