Activated the Hand-on Menu + Added Name of feature to icon
Now, when you start the project, the menu will be automatically activated. Also, for the feature name, I successfully added the feature name from the AddFeatureFromServer() within the FeatureManager.cs. So, now when you change the name of the features on the server, the icon name will change as well!