Map Icon Spawn Updates
Using the new map image we've updated the spawn location and the prefabs. Now the axis for the map spawn parent has z is "North" where the top of the map is "NOrth". I removed some of the other inversions we were doing. Now the map icons have the same coordinates as their server counterparts. There is still a bit of uncertainty over the location of the origin. for some reason the current method only works if I multiply the x,z coordinates by -1, however, that can be addressed later. For now it seems to be spawning in the correct location. the prefabs had to be rotated 180 in the x axis to prevent back face culling