Upgraded the rotation script, only y rotaiton
Now we can control which rotation components we copy from the source object. In this case we only want the compass orientation, ie y axis rotation, to copy over. The other axis will cause the map icon to clip into the map.
TODO: Only rotate the local headset icon based on the camera value and not all of the icons