Map updating size and origin!
Now the map should update it's aspect ratio and size each time the hand map activates. I created a placeholder MapController that brings the texture request for the map image from the EasyVizArHeadsetManager into the controller class. I've also started work on depreciating and migrating the MapIconSpawn class that was not doing anything with spawning, but rather was setting the aspect ratio. There's a current dependacy with the navigation class looking for that script, so I haven't been able to completely remove it yet, but I'll work on getting navigation updated too.
The map controller isn't really wired up fully yet. It needs to get an event from the map visuals that they are active, but currently it's attached to the map game object and updates on the OnEnable() monobehavior call.
I also tried generating some class diagrams