Map Navigation
I think we have on map navigation now! I need to test it on the headset with references to see if the path is correct, but it updates on new target presses. Maybe I've broken some of the target selection logic. Now it doesn't compose the correct JSON for the server. I think it's because the target variable isn't updated correctly? I think maybe it's working?? It maybe was server lag or too many button presses?