Updating the destination position
We need to patch the destination for the path to the server. I was getting patch errors that I thought had to do with the headset target data. Turns out it has to do with the Local headset ID not being set. This is due to some logic in the creation of the headsets that isn't correctly updating the boolean value that flags if a headset is local