Boolean cleanup and Bughunt
There's a problem in the Local headset logic. No headset is ever marked as local and the registration doesn't save because there's a "Received an error when creating headset" from the CreateCallback() method.
There were also 3 "is local" type variables in the headset manager class, so I tried to refactor them into a unified is_local field.
I've hit a bit of a wall when it comes to the headset creation logic. There's something that's preventing the correct local headset id from being saved. the program seems to think that the local headset is "EasyVizarHeadset(Clone)" but that's not a fully created headset and has no ID.