Debugging spheres
It turns out testing this was a bit trickier than expected. I wasn't able to get the raycast path to show up on the headset. I also tried adding some spheres into the mix to see if we could at least see those. They do show up, but they are taking their height from the headset's height, so that's not quite right yet,