Updated navigation path
I found there was a line render call in the navigating path callback. So, everytime the path was retrieved another call to modify the path was being made that I was not aware of. I disabled that call because the retrieval of the data should be seperate from it's visualization. I also had a allocation size error that was expanding the array after trying to add the entry. This was causing a index out of bounds error in some cases as well.