Added Asset Search and Working on distance calc code
I was trying to fix the patch error, but tracing it through the code I found something is modifying a public variable that holds the headset ID that is causing the patch error. We need to get the correct headset ID and remove the access to these vairables from other classes.
This lead me to the distance calculation code, which has some map icon errors. The when you first move the distances are incorrect, but when you stop moving they update to the ocrrect value. The worldspace icons are correct, so instead of calculating the distances and map features seperately I'm working on only calculating that value once and having the views of the icons accesss a common data structure.
I added in a new unity asset that works much better than the find references in scene menu option. This can actually find the game objects and prefabs the scripts are attached to instead of the deafulat Unity behavior. I'm not really sure how the built in Unity one works because it usually doesn't return the results I'm expecting.