Removing cur_prefab
Through the debugging tools and tracing I couldn't quite find where the cur_prefab was being set or modified. Even when the breakpoints were at the spawning of the headset they would be initialized before I could see what was doing it. I looked at what the variable was used for and it seemed to be a reference to the game object holding the script. I swapped in a temp this.gameObject and everything worked! This showed me I could fix the problem. I wasn't satisifed because there was still something else "out there" setting the value. Eventually I looked for the headset prefab and found the distance calculation script there. Inside the prefab the cur_prefab field was set to the contaning prefab itself! Confirming my this reference suspicion. I also saw that when the variable name is changed, the editor inspector reference was lost. This made sense to the null reference error we were seeing before. After finding this I set the variable in the script instead of the editor and made it private. It's still working as expected. My next step is to remove the references to that variable alltogether and use this.gameObject in it's place.