Fast script reload and Map icon error fix
There was an error with a null reference in the map icon spawning. We didn't notice it before because it only affects one of the prefabs. It was the headset prefab. When Unity pulls the features from the server it would throw a null reference error when trying to spawn the headset icon. I checked the dictionary and key values and names and those looked good. Eventually I found it in the color changing script. The script was looking for a "quad" game object child and trying to get it's renderer component. In the headset icon the component with the render was called "Icon Visuals" so it was returning null when the color was changed and stopped the spawning of the other icons. I updated the code so that all of the prefabs now have "Icon Visuals" instead of "Quad" game objects for icon rendering. I also made the implicit rotations for the renderer, colliders, and text fields consistent so that they are now all "down looking" buttons. This way, the local axis on the map are mirrored in the real world coordinate system. Now we don't have to do any rotations when pulling the data from the world space icon transforms.

I added another asset called FastScriptReload. It's free from the asset store. It requires you to change the Unity project settings for asset reloading. I also included a screenshot of the warning insturcionts in case we want to remove it. It's been working well so far with much faster script reloads, though I do think it has made Unity slightly less stable. I had it hang twice during today's work session. However, the time it's been saving in the debugging process seems worth the amount of time it costs to relaunch the Unity project. If it causes trouble we can always take it out later.