Local Headset tracking
I'm narrowing in on the local headset identificaiton an manipulation. I found out that there was a field that needs to be set in a new headset before the JSON data is copied into it. it's the
headset_class_data.headset_parent
field. If this is not set before one calls
            headset_class_data.AssignValuesFromJson(local_headset);
a null reference will occur.
I think we need to figure out a better way to fix this than having code dependent on ordering, but I've got it at least working again after hitting that wall. We should be getting less warnings now too.