Animated Headset Icon and Breaking apart Distance calc
I got things back to the point where the local headset is spawned on the map. I was having some rather confounding null reference errors with a find call.
It was supposed to test if the current local headset icon had been spawned, but despite it looking okay in the debugger was gettin null ref errors. I eventually figured out it was a transform.Find() call and not a gameObject.Find() call. this returns the transform of the matching object, or null if not found. In this case we were using a transform.gameObjectt reference. However, when the transform = null. a null.gameObject call produces an error. It was working before because of the combination of the distance calculation, animation, and spawning methods. However, when I started pulling these apart there was a local variable that went out of scope and that caused this error to surface.