Constructor, but BROKEN Code
I got some constructors in and working, but there was a change that I made that is causing a null reference. there is a check against a string comparison, but the string is null. I need to get the constructor finished before this can be fixed though I think. revert to previous check in if you need to deploy