REPAIR The constructors are fucntional
I've fixed the null reference errors and am now working on getting the constructors in use. I will have to remove the EasyVizARHeadset script from the prefab. I am going to add it via scripting at run time because it's data is overwritten by the JSON when updated, so it's not as useful to have as an editor accessible thing right now.
I have to use the add at run time because I can't modify the values in place on the game  object, though I suppose I could try and create an update function instead. Either way. things are working right now, so this is the checkpoint for REPAIRed work