Headsetup
I'm really struggling with the callbacks. they've made this really tricky to do what I'm trying to do. I also don't konw If I can parse a list vs a headset singular. The get call from the server on this returns both a single headset, or a list, depending on the ID. but if i'm trying to see if the ID is registered based on the retunred data I need to know if the ID is valid in the first place.....