IP Address and Spawning cleanup
We had to switch from the URL to an IP address for the EasyViz server because of a DNS server somewhere on the UW/WID network blocking connections. The IP address of the server is 128.105.22.228:5000, and for  the QR code URL it needs to be prefixed with the vizar://
Websocket will also work with IP address.
Removing an extra call to the create headset function that ran before the QR was registered.
Removed a Unity editor if declaration that was using a hard coded server and location ID instaed of the data passed by the QR scan