Moving the feature manager into a feature manager
In looking at some of the Unity errors I found that there were 5 instances of the feature manager script and that the one being used was on a gameobject that was going to be deleted.
I am getting errors on the posting of new features, but I'm not sure if that's been like this for a while or if I just broke things. Going to do a revert to check.