Fixed the registration bugs!
I think I've resolved the issue with the registration. Now the app will recognize the registration and correctly check against the server. If the registration file is deleted it will get a new one and set the Local ID to that value. If the registration is deleted from the server it will request a new one. I'm now looking at the position updates as those don't seem to be sending to the server.