Position Posting Problems
The post position code is never reached. There's a bunch of boolean checks that have ambigious purposes and the time delay funciton in the update might not be set up correctly. I'm going to first move it to the co-routine and see if I can also eliminate some of the extraneous variables.