Patching is working with the server now!
Position patching is up and running, uses the headset prefab's update interval, but that has to be set before the application starts because the co-routine is launched in the start function.  Cleaning up some of the headset functions and variables that don't need to be used. Lance is working on getting the websockets working correctly so we can move the posting and location updating from that in the future.