HMD doesn't register, but editor does??
Tested the new code on the HMD and I'm not getting any registration from the HMD. The editor works fine and handles all the cases, but the HMD won't register and I can't tell why yet.
I wonder if it has to do with
EasyVizARServer.Instance.SetBaseURL(base_url);
call in the QR reader. If there's no local registration this will fail, and even if a regsitration is recieved, then saved, it might be causing some sort of problem on the HMD. If I remove the _hasRegistration = true in the save registration function, then it gets stuck in an infinite loop, but I think that should be there. Maybe I'm using it wrong though. I think I need Lance's help to understand this.