No holohands and feature spawn retool
We've removed the holo hand overlay in the HMD, though it should still be visible in the editor. Also re-working how the features are spawned and updated. It seems like there's a caching behavior going on, and deleting all the features at the start of the app might help. I'm also putting the Websocket update rate down to .5 seconds as there seems to be significant latency when uploading the map meshes, and I'm not sure what's causing that. Maybe the Websocket requests are filling the queue too fast?