Mult-layer Map setup
Working on getting a multilayer map setup to use for multiple floor locations. Some functions still need to be implemented, but the main logic is sketched out now. Have some quesitons on how to pass additional arguments to a callback function.
Also a TODO, the line rendered path on the map for navigation should zero out the y values of the line because they are dipping below the surface of the map and hiding.