Fire Eyes Faster
I setup a GPU instanced material for the eye tracking painting and it's much faster on the laptop. Haven't checked on the HMD solo yet. Also, I was able to increase the spatial observer mesh side to a 30x30m volume with higher quality meshes. Idk how that will affect performance, but it was looking nice on the HMD and solves the issue of not being able to paint a wall that was too far away. before the volume was 3x3m