Map Clipping Setup
I have the spherical map clipping working finally! It was a little tricky, but it seems that the instanced materials clipping was causing some problems with the way that we assign the PNG perhaps? Either way I shifted it over to the shared material modificaiton and it seems to be working. I don't think there's any problem with using the shared material because we're going to want this cropped for all the hand maps. This could cause an issue for a worldspace map though, so those will probably need their own material so as not to be affected. The next steps are:
~figure out how to center the user in the middle of the map view.
~clip out the out of bounds icons
~decide what is a reasonable size for the map view