Icon clipping too!
I added a script to manage the icons that are on the clipping map. Now they are hidden when outside of the map radius. This assumes a spherical clipping volume. I initially was using a bound.intersection set of calls with a collider, but was getting strange results. There would be objects that should have been outside of the sphere collider, but they were still active. I eventually deduced that this was due to Bounds being a cube! thus because my visual was a sphere, I was getting an area that I wasn't expecting that should intersect with the icons. I eventually had a realization that I could use a Vector3.distance based approach and that worked the way I was hoping it would as well as being a more simplified approach