Activation and deactivation of componenets
I've set up the componenets to properly activate and deactivate based on the adjustement state of the map. The box colliders need to be toggled between the two grabbale objects because if one collier is inside the other only the outer one is activated. I've also disabled the mesh renderer for the bounding box when it's not active.