Updated the spawning and movement
Now the bottom of the map spawns where the globe is placed, making it easier to adjust the buttons. Unfortunately the button placement code is kinda broken. in the MRTK implementation the y offset isn't used. i was able to make a fix, but the script is stored in the packages part of Unity so it's not synced via github. I'm not quite sure what to do with that. We could make our own AppBar.cs copy, but I'm not sure what the consequcnces will be. I wasn't able to figure out how to make the box immedeatly editable yet, it still needs the button press at the moment. I also am seeing some weird distance interaction behavior where the cursor collides with a box collider that shouldn't be there anymore.