Smaerter Volume Map
Changed some of the logic to have the bottom of the mesh be at the same location as the bottom of the culling reagoin. Now when the cull volume is locked in the map will be at the base of it. this was tricky because there was an unexpected modification of the BoxCollider on the moveable map by the bounds control component from the MRTK. It acutomatically tries to fix the associated box collider to the first mesh renderer that it has in it's hierarchy. The problem here was that this box collider was calculated before I moved the map to be at the floor of the culling region, so it was above the locaiton of the map. I've added the floor adjustment to the awake function and that seems to execute before the bounds are checked. We will have to keep this in mind if we update the model from a new source like the server. Also, it might not always be desired to "floor" the model after adjusting the bounds, so I've added a conditional funciton call that will only adjust the bounds once and then a boolean is flipped. This can be updated publically, so ut should allow us to make changes later. I envison the scenario where you wouldn't want the auto flooring is when you have a multi floor map and you adjust the bounds, but whant to keep your place. This might not be a reasonable use case, but the behavior is kinda undefined at the moment.