Clipping Enabled!
Now we have clipping enabled. I had to remove the material instanced option and switch to shared materials for this to work correctly. The full solution is in the VLC branch for further explanation. It should be working for all the scenes too because I made the necesary edits to the prefab itself, so it should propagate to older scenes too as long as they have the "Palm Up Map"
Added a different set of colors to the clipped map.
The button to the right of the map will now cycle between the colors and the clipped and non clipped maps.
Enabled the user offset centering. It isn't quite in the center of the map, but it moves with the user as they explore the envrioment. I think I need to move the culling sphere as well because it should be centered on where the user is, but right now it's on the center of the map itself.