Changed the sign behavior for better usability
the object that you place in the world was still a child of the tool menu, so if the other hand went out of tracking range then the active object gets stuck. now when you grab the item from the dock it will remove itself from the hierarchy and so only the hand holding the object needs to maintain tracking.