Disabled the map clipping for the demo
The clipping and user centered offset were looking weird with this map shape, so I disabled the components.
I'm not sure if this was due to the map shape itself being smaller and wider than the more squared version I was testing on the downstairs lab. It may also be an issue with how we're centering the map. This will be looked at later, but the biggest takeaway is that we should probably have a way for the user to turn on and off clipping so that they can see the entire map.