Sign Placement Integration
Adding the sign placement to the grab and drop gestures for the dynamic sign.
If the QR scanner is scanned before the signs are placed this will likely break signs, but that is not the intended usage pattern and would be user error.
The signs look to update their arrows based on the camera location.

We may need to update the manipulation events to trigger a sign update based on where the sign is actually placed.