Remove Duplicate Nav Manager and Fix Arrow
The nex player UI elements were being used in the sign as the arrow, so wehn they were removed the sign's material broke. I restored the necessary files and will eliminate more specifically the compiling error files so we can still use the arrows. I tested to make sure the arrows are appearing again and they are.