Dockgrab Cloner Encapsulated
Moving some of the public variable calls and usage from the volume maps so that the nav mesh update doesn't use as many find calls.
The navmesh map seems to be currently hardcoded to reference the Example Dev Lab Unlimited file.
I am not getting the memory leak from the loading of the server maps.