Added a new inspector utility
Added a new package to allow for the viewing of dictionaries in the Unity editor. You must use the VInspector SerializedDictonary to see them

Finished the navigation path visualizer. You can set the starting point of the visualizer and specifiy one or all features in the locaiton to be navigated to.

I tried to keep it general so we should be able to use it for the world space, the map space, the volume space, and sign space line paths. This way we have a line renderer per feature and path.

Not fully implemented yet, there still needs to be some data fields that I update.

Improved the sign functionality by zeroing out some of the offsets that took the object's center away from 000.