Added a pathfinding offset to the sign
The paths the sign displays are now able to be assigned a starting point for the pathfinding. If a value is not assigned it will use the script's locaiton. Currently there's an offset on the sign that we will use as a astarting locaiton. It's 15cm from the wall right now, but can be adjsuted.