Trying to account for degrees
Working on accounting for the degrees of sign rotation. I think the logic is that we want to undo the rotation the sign has with respect to the origin. If we do that, then the coordinates from the navigation mesh will be in the same rotation space. That mean's we just need to apply that rotation before we intreperate the direction vector between the two points. that does the rotating to the signs coordinate space so that the arrow can be interperated off that angle