Bounds tricky
I think I figured out the issue with the bounds. They aren't calculated with respect to new transform positions perhaps?
https://docs.unity3d.com/ScriptReference/Renderer-bounds.html
When you set custom world bounds, the renderer bounding volume no longer automatically tracks Transform component changes.
Maybe this is it? Or also, it might be that we need to zero out the server map transform position every time we add the offset. I think this was looked at for the boolean. STill, this is quite a bit more confusing than expected, but I might have just been staring at this for too long today lol.