Better behaved bounds
Now! We have the bottom matching correctly I believe. I also have added an autoshrink function to the 3D meshes so that they will be scaled to the smallest factor (X,Y,Z) that allows the map to be fully displayed within the bounding volume :D
I had to take the smallest axis and apply it uniformly because if I tired to scale on a per axis encapsulation it would skew and distort the model.