More robust docked apps
We've had to keep both hands in view of the hololens as we pick and place docked tool objects on the hands. However, I think I can make it so that we only need to use the hand that's holding the object now. I've removed the child objects from the hand hierarchy, so now, if the hand loses tracking it doesn't also disable the object because it's no longer one of the children!