I changed to R/W allowed
I don't know if this helped or if it was the clean build, but the icon did finally showup! but alas, it's really small so you can't really see much, but it is blue and red so at least that has some first responder conntations. I might try and update it later, but IDK which one of these images actaully is being used, so it might no be worth the effort to change it.