Fixed the non deleting globe
I tried to change the implementation to fix the non-destrtucting globe, but the real fix turned out to be changing the order of operations of the events in the on manipulation end part of the editor settings. I think adding a invoked destruction might have worked as well, but my guess is something about executing the manipulation handling somehow stops the destroy command from working?