Sort of vetex coloring!
We have a CustomOBJ loader that you can point at the streaming assets frolder with the source OBJ in it, and it will spawn the vertex color parsed version of the model. You then need to give it an appropriate vertex color shader, but there's a custom shader in here now as well as a material that's been set up for vertex shading. not sure if it's going to work in the runtime import process, but it's a start.