Setting up the ghost grabber
I had to rework the volume map avatar a bit so that we can better interact with it. I've got a ghost placeholder now. the initial behavior is looking good. I did manage to break the path preview again lol, but i probably won't be able to totally fix it until i get the rest of the avatar interactions finished.
you can grab at the user's current position to spawn the ghost cheese. it will stay where you leave it until you try and set a new path for the user by grabbing on them again.