The marker menu is now server controlled on startup
We take the setting from the server config of the headset for showing the user the maker menu after QR scanning.

I was able to get around the not finding inactive objects because the content that is being toggled is actually a child of the main object, so I set up a function to feed a bool into that will control that componenet.

I want to move the toggle button onto the wrist menus so that if we want to summon the markers we still can.