Open XR Error
There was an error I was getting during runtime that was not crashing the app but causing some problems. it was only on
my computer though.

Failure to restart OpenXRLoader after shutdown.
UnityEngine.SetupCoroutine:InvokeMoveNext (System.Collections.IEnumerator,intptr)

Eventually I found a setting in Unity that set the OpenXR to the system default, but when I switched it's playmode OpenXR runtime to Mixed Reality Toolkit it made the errors go away. My guess is my system Open XR runtime might not be that and so it was launching the wrong software.