Updaing drone visual for proper size and center
The drone orbs are set with their center touching the ground, but we want that to be in the middle. I've updated the offsets so they now are have the drone in the center.
Drones will also have a different spawning behavior for the visuals to take their position from the center of the data point instead of offset like hmds