Drone visuals mode
Added a mode for drone visuals to be centered on the drone, but I put it in the cycle script for the enabling, and so the first visual wouldn't show up correctly. I moved that to the headset type boolean chain so now it should be good to go from the get go.